corrections according to the review

+++ src/main/java/ru/yandex/practicum/filmorate/storage/mpa/MpaStorage.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface MpaStorage {
 
     List<Mpa> getMpas();
-    Optional<Mpa> getMpaById(int id);
+    Optional<Mpa> getMpaById(long id);
 }
